+++ R/build.r
@@ -244,6 +244,9 @@
 #' @inheritParams build_reference
 #' @param lazy If `TRUE`, will only rebuild articles and reference pages
 #'   if the source is newer than the destination.
+#' @param new_process If `TRUE`, will run `build_site()` in a separate process.
+#'   This enhances reproducibility by ensuring nothing that you have loaded
+#'   in the current process affects the build process.
 #' @export
 #' @examples
 #' \dontrun{
@@ -252,6 +255,75 @@
 #' build_site(override = list(destination = tempdir()))
 #' }
 build_site <- function(pkg = ".",
+                       examples = TRUE,
+                       document = TRUE,
+                       run_dont_run = FALSE,
+                       seed = 1014,
+                       mathjax = TRUE,
+                       lazy = FALSE,
+                       override = list(),
+                       preview = NA,
+                       new_process = TRUE) {
+
+  if (new_process) {
+    build_site_external(
+      pkg = pkg,
+      examples = examples,
+      document = document,
+      run_dont_run = run_dont_run,
+      seed = seed,
+      mathjax = mathjax,
+      lazy = lazy,
+      override = override,
+      preview = preview
+    )
+  } else {
+    build_site_local(
+      pkg = pkg,
+      examples = examples,
+      document = document,
+      run_dont_run = run_dont_run,
+      seed = seed,
+      mathjax = mathjax,
+      lazy = lazy,
+      override = override,
+      preview = preview
+    )
+  }
+}
+
+build_site_external <- function(pkg = ".",
+                                examples = TRUE,
+                                document = TRUE,
+                                run_dont_run = FALSE,
+                                seed = 1014,
+                                mathjax = TRUE,
+                                lazy = FALSE,
+                                override = list(),
+                                preview = NA) {
+  args <- list(
+    pkg = pkg,
+    examples = examples,
+    document = document,
+    run_dont_run = run_dont_run,
+    seed = seed,
+    mathjax = mathjax,
+    lazy = lazy,
+    override = override,
+    preview = FALSE,
+    new_process = FALSE
+  )
+  callr::r(
+    function(...) pkgdown::build_site(...),
+    args = args,
+    show = TRUE
+  )
+
+  preview_site(pkg, preview = preview)
+  invisible()
+}
+
+build_site_local <- function(pkg = ".",
                        examples = TRUE,
                        document = TRUE,
                        run_dont_run = FALSE,
@@ -285,16 +357,6 @@ build_site <- function(pkg = ".",
   build_tutorials(pkg, override = override, preview = FALSE)
   build_news(pkg, override = override, preview = FALSE)
 
-  preview_site(pkg, preview = preview)
   rule("DONE", line = 2)
-}
-
-build_site_rstudio <- function(pkg = ".") {
-  callr::r(
-    function(...) pkgdown::build_site(...),
-    args = list(pkg = pkg),
-    show = TRUE
-  )
-  preview_site(pkg)
-  invisible()
+  preview_site(pkg, preview = preview)
 }